requirements.txt
pyrogram
tgcrypto
pymongo[srv]
python-dotenv
